Move DocSystem tests to a sep. file, use testsets
